Fixed calling vsnprintf.
git-svn-id: svn://svn.code.sf.net/p/pspdftool/code/trunk@21 bcc72368-3e3e-0410-9a69-86e283d76c0a

+++ ps_lib.c
@@ -708,7 +708,7 @@ static void getPname(char * kam, const char * odkud){
 	while (odkud[i] && odkud[i]!=':')
 		++i;
 	++i;
-	while (odkud[i] && isspace(odkud[i]))
+	while (odkud[i] && isspace((int)(odkud[i])))
 		++i;
 
 	if (odkud[i]=='('){
@@ -720,7 +720,7 @@ static void getPname(char * kam, const char * odkud){
 		kam[j]=0;
 	}
 	else {
-		while (odkud[i] && !isspace(odkud[i])){
+		while (odkud[i] && !isspace((int)(odkud[i]))){
 			kam[j]=odkud[i];
 			++i;++j;
 		}
